feat(client): add failover mechanism
Go through the entire list of connector in case of hard failure
(e.g. when the connector can't be accessed or is timing out,
not when the request is reportedly bad (e.g. get request with
keys that are non-existant))

This required to append a value to the Error instead of just sending
a `new Error()`. This allows us to differentiate hard failures
from expected failures.

Updated the unit tests accordingly.