fix(keygen): use safer entropy for ARC paths
node-uks was relying on `Math.random()` which would create collisions in
keys way too often because of its lack of entropy. We now have a simple
function creating a lookable key in the system with enough entropy.

The entropy is effectively 88 bits, i.e. 11 bytes.

Tests are updated to use a mockup server.

BREAKING CHANGE: The put prototype has been updated to allow keygen
based on some parameters.