bugfix: SPRXCLT-5 do not abort PUT requests
Fix behavior with node 16 by not relying on the emission of the
'close' event of the source stream to decide that the request to
sproxyd should be aborted. Instead, use the "stream.finished" helper
function that is designed to handle this case.

It is believed that the regression comes from a change in the default
value of "autoDestroy" on streams to "true" in node 14.x, leading to
streams being destroyed automatically after completion.